Move read_geoparquet_async and parse_geoparquet_metadata functions
Create a dedicated reader_async.rs file to put read_geoparquet_async, and move the parse_geoparquet_metadata function from reader.rs to geoparquet_metadata.rs.
