+++ src/io/parquet/mod.rs
@@ -2,5 +2,9 @@
 
 mod geoparquet_metadata;
 mod reader;
+#[cfg(feature = "parquet_async")]
+mod reader_async;
 
-pub use reader::{read_geoparquet, read_geoparquet_async, GeoParquetReaderOptions};
+pub use reader::{read_geoparquet, GeoParquetReaderOptions};
+#[cfg(feature = "parquet_async")]
+pub use reader_async::read_geoparquet_async;
